requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
cssselect==1.3.0
distro==1.9.0
fastapi==0.115.14
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.5.1
google==3.0.0
google-auth==2.40.3
google-genai==1.26.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.1
openai==1.95.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
text-unidecode==1.3
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
